refactor: Service클래스 Interface 분리
