+++ src/main/java/lotto/config/AppConfig.java
@@ -1,6 +1,8 @@
 package lotto.config;
 
 import lotto.controller.LottoController;
+import lotto.service.LottoService;
+import lotto.service.LottoServiceImpl;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
@@ -15,6 +17,10 @@ public class AppConfig {
     }
 
     public static LottoController lottoController() {
-        return new LottoController(inputView(), outputView());
+        return new LottoController(inputView(), outputView(), lottoService());
+    }
+
+    public static LottoService lottoService() {
+        return new LottoServiceImpl();
     }
 }
